updated README | added icons to tabs | moved evolution out of tabs | added missingno to 404 page
