+++ app/(tabs)/_layout.tsx
@@ -5,6 +5,7 @@ import { Platform } from 'react-native';
 import { HapticTab } from '@/components/HapticTab';
 import { IconSymbol } from '@/components/ui/IconSymbol';
 import TabBarBackground from '@/components/ui/TabBarBackground';
+import AntDesign from '@expo/vector-icons/AntDesign';
 import { Colors } from '@/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
 
@@ -31,6 +32,20 @@ export default function TabLayout() {
         options={{
           title: 'Home',
           tabBarIcon: ({ color }) => <IconSymbol size={28} name="house.fill" color={color} />,
+        }}
+      />
+      <Tabs.Screen
+        name="Search"
+        options={{
+          title: 'Search',
+          tabBarIcon: ({ color }) => <AntDesign name="search1" size={24} color={color} />,
+        }}
+      />
+      <Tabs.Screen
+        name="RandomPoki"
+        options={{
+          title: 'Random',
+          tabBarIcon: ({ color }) => <AntDesign name="question" size={24} color={color} />,
         }}
       />
     </Tabs>
